AB#73129 Fix for parse-link-header not working
Under vite the builtin NodeJS modules aren't available. parse-link-header uses the URL module and this isn't available, adding a polyfill is rather messy. So instead we switch to a more up to date package that removes the dependency on the node modules.

Here's where the fix was basically rejected in the original package: https://github.com/thlorenz/parse-link-header/issues/23